third_party/WebKit/LayoutTests/platform/win/http/tests/security/contentSecurityPolicy/require-sri-for/require-sri-for-worker-fromblob-blocked-expected.txt

CONSOLE ERROR: line 9: Refused to load the script 'blob:http://127.0.0.1:8000/534d99c9-61ef-4b85-bd41-5e56d5d9a7e3' because 'require-sri-for' directive requires integrity attribute be present for all scripts.
This is a testharness.js-based test.
PASS Test that a worker can be created. 
Harness: the test ran to completion.
